refactor and fix some props
